refactor dict package
