+++ libakaza/src/dict/mod.rs
@@ -1,3 +1,3 @@
 pub mod loader;
 pub mod merge_dict;
-pub mod write;
+pub mod skk;
