+++ libakaza/src/dict/skk/mod.rs
@@ -1,0 +1,3 @@
+pub mod ari2nasi;
+pub mod read;
+pub mod write;
