--- libakaza/src/skk/mod.rs
@@ -1,2 +1,0 @@
-pub mod ari2nasi;
-pub mod skkdict;
